Bug 1198978 - Make sure the Save button doesn't do anything if the pinboard is empty
Previously, if you entered things into the bug or classification fields
and then hit the "Save" button without first actually pinning a job to the
pinboard, your bug/classification information was wiped away because the
save function only verified that you were logged in, not that there was
anything in the pinboard that needed saved.

The save function now checks that there's something pinned as well before
continuing on with the save function.