Bug 1524895 - Prompt for retrigger 'times' if shift key is held
This adds a new modal named RetriggerMultiplePrompt which contains a single numbers-only input. When the modal form is submitted, the value of that input is used as the "times" property when calling taskcluster's "retrigger" action task. This allows for multiple retriggers of one or more jobs to be created from a single action task, which will hopefully significantly cut down on the number of action tasks spawned while tracking down failures, since the pool of workers that can process action tasks is somewhat limited.

The modal is applied to the actionbar's retrigger button and the pinboard's "retrigger all" button, as well as adding a keyboard shortcut (shift-r).